requirements.txt
# Core dependencies
langchain-core>=0.1.0
langchain-mcp-adapters>=0.0.3
langgraph>=0.0.26

# Language models
langchain-openai>=0.0.1
langchain-google-genai>=0.0.1
langchain-anthropic>=0.1.1
langchain-groq>=0.0.1

# Server and networking
fastapi>=0.104.1
uvicorn>=0.25.0
httpx>=0.25.0
aiohttp>=3.9.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1